+++ src/main/java/com/mysite/sbb/Repository/CommentRepository.java
@@ -3,6 +3,6 @@ package com.mysite.sbb.Repository;
 import com.mysite.sbb.Model.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommentRepository extends JpaRepository<Comment, Integer> {
+public interface CommentRepository extends JpaRepository<Comment, Long> {
 
 }
