+++ src/main/java/com/mysite/sbb/Repository/QuestionRepository.java
@@ -12,7 +12,7 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 import java.util.Optional;
 
-public interface QuestionRepository extends JpaRepository<Question, Integer> {
+public interface QuestionRepository extends JpaRepository<Question, Long> {
 
     // findBy + 엔티티의 속성명
     QuestionDTO findBySubject(String subject);
@@ -21,7 +21,7 @@ public interface QuestionRepository extends JpaRepository<Question, Integer> {
     // N+1 문제를 해결하기 위함
     // 그냥 join fetch로 하니 질문에 답변이 없으면 null을 뱉어버림. 그래서 left 추가.
     @Query("select q from Question q left join fetch q.answerList where :id = q.id")
-    Optional<Question> findById(@Param(value = "id") Integer id);
+    Optional<Question> findById(@Param(value = "id") Long id);
     // 제목에 특정 문자열이 포함되어 있는 데이터 조회.
     List<Question> findBySubjectLike(String subject);
     Page<Question> findAll(Pageable pageable);
