+++ src/main/java/com/mysite/sbb/Service/AnswerService.java
@@ -9,8 +9,8 @@ import com.mysite.sbb.Model.SiteUser;
 
 public interface AnswerService {
     AnswerDTO create(QuestionDTO questionDTO, String content, SiteUserDTO author);
-    AnswerDTO getAnswer(Integer id);
-    Integer getQuestionId(Integer id);
+    AnswerDTO getAnswer(Long id);
+    Long getQuestionId(Long id);
     void modify(AnswerDTO answerDTO, String content);
     void delete(AnswerDTO answerDTO);
     void vote(AnswerDTO answerDTO, SiteUserDTO siteUserDTO);
