+++ src/main/java/com/mysite/sbb/Service/CommentService.java
@@ -7,7 +7,7 @@ import com.mysite.sbb.Model.Answer;
 
 public interface CommentService {
     void create(AnswerDTO answerDTO, String content, SiteUserDTO siteUserDTO);
-    CommentDTO getComment(Integer id);
+    CommentDTO getComment(Long id);
     void modify(CommentDTO commentDTO, String content);
     void delete(CommentDTO commentDTO);
     void vote(CommentDTO commentDTO, SiteUserDTO siteUserDTO);
